IBAN national check digit support
Some countries include a national check digit in their BBAN. If this
check digit is incorrect, the derived IBAN is considered invalid.

This change begins the process of supporting national check digits.
Because each country may have their own implementation, this change
focuses on the ISO 7064 MOD-97-10 algorithm and its variants.

I've included configuration for France, Portugal, and Serbia - all of
which use the ISO 7064 MOD-97-10 algorithm. France uses their own
variation of this algorithm in which the BBAN letters are encoded
differently.